rlhf/requirements.txt
accelerate==0.18.0
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
boltons==23.0.0
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.3
comm==0.1.3
conda-build==3.23.3
conda-package-handling==1.9.0
conda==23.3.1
cryptography==38.0.1
datasets==2.11.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.6
dnspython==2.2.1
docker-pycreds==0.4.0
exceptiongroup==1.0.4
executing==0.8.3
expecttest==0.1.4
filelock==3.6.0
flit-core==3.6.0
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.2
gitdb==4.0.10
gitpython==3.1.31
glob2==0.7
huggingface-hub==0.14.1
hypothesis==6.61.0
idna==3.4
ipykernel==6.22.0
ipython==8.7.0
jedi==0.18.1
jinja2==2.11.3
jsonpatch==1.32
jsonpointer==2.1
jupyter-client==8.2.0
jupyter-core==5.3.0
libarchive-c==2.9
markupsafe==2.0.1
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
nest-asyncio==1.5.6
numpy==1.22.3
packaging==23.0
pandas==2.0.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.3.0
pip==23.0.1
pkginfo==1.8.3
platformdirs==3.3.0
pluggy==1.0.0
prompt-toolkit==3.0.20
protobuf==4.22.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycosat==0.6.4
pycparser==2.21
pygments==2.11.2
pyopenssl==22.0.0
pysocks==1.7.1
python-dateutil==2.8.2
python-etcd==0.4.5
pytz==2022.1
pyyaml==6.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.1
responses==0.18.0
ruamel.yaml.clib==0.2.6
ruamel.yaml==0.17.21
sentry-sdk==1.21.0
setproctitle==1.3.2
setuptools==65.5.0
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.2.0
sympy==1.11.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch
torchelastic==0.2.2
torchtext==0.14.1
torchvision==0.14.1
tornado==6.3.1
tqdm==4.64.1
traitlets==5.7.1
transformers==4.28.1
trl==0.4.1
types-dataclasses==0.6.6
typing-extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wandb==0.15.0
wcwidth==0.2.5
wheel==0.37.1
xxhash==3.2.0
yarl==1.9.2